feat(refactor): move query support classes into plugin apis
