+++ packages/data/src/exported.ts
@@ -4,6 +4,3 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
-export { QueryOptions, SoqlQuery, SoqlQueryResult } from './soqlQuery';
-export { Field, FunctionField, SubqueryField } from './queryFields';
