+++ packages/data/test/placeholder.test.ts
@@ -4,5 +4,3 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export * from './testUtil';
-export * from './soqlQuery.exemplars';
