+++ packages/plugin-data/test/api/data/soql/test-files/queryFields.exemplars.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { Field, FunctionField, SubqueryField } from '../lib/queryFields';
+import { Field, FunctionField, SubqueryField } from '../../../../../src/api/data/soql/queryFields';
 
 export const makeSubfield = (name: string, fields: Field[]): SubqueryField => {
   const subfield = new SubqueryField(name);
